move selectors in separate files
