+++ src/pages/forgot/forgot.jsx
@@ -5,7 +5,8 @@ import { useDispatch, useSelector } from 'react-redux';
 
 import { routes } from 'routes';
 
-import { userSelectors, userActions } from 'resources/user/user.slice';
+import * as userSelectors from 'resources/user/user.selectors';
+import { userActions } from 'resources/user/user.slice';
 
 import Input from 'components/input';
 import Button from 'components/button';
