+++ src/pages/reset/reset.jsx
@@ -4,7 +4,8 @@ import { Redirect } from 'react-router-dom';
 
 import { routes } from 'routes';
 
-import { userSelectors, userActions } from 'resources/user/user.slice';
+import * as userSelectors from 'resources/user/user.selectors';
+import { userActions } from 'resources/user/user.slice';
 
 import Input from 'components/input';
 import Button from 'components/button';
