+++ src/resources/toast/toast.selectors.js
@@ -1,0 +1,1 @@
+export const selectMessages = ({ toast }) => toast.messages;
