+++ src/resources/user/user.selectors.js
@@ -1,0 +1,1 @@
+export const selectUser = ({ user }) => user;
